+++ guns-admin/src/main/java/cn/enilu/guns/admin/modular/system/controller/LogController.java
@@ -70,7 +70,7 @@ public class LogController extends BaseController {
     @RequestMapping("/detail/{id}")
     @Permission(Const.ADMIN_NAME)
     @ResponseBody
-    public Object detail(@PathVariable Integer id) {
+    public Object detail(@PathVariable Long id) {
         OperationLog operationLog = operationLogRepository.findOne(id);
         Map<String, Object> stringObjectMap = BeanKit.beanToMap(operationLog);
         return super.warpObject(new LogWarpper(stringObjectMap));
