+++ guns-admin/src/main/java/cn/enilu/guns/admin/modular/system/controller/LoginController.java
@@ -51,7 +51,7 @@ public class LoginController extends BaseController {
     @RequestMapping(value = "/", method = RequestMethod.GET)
     public String index(Model model) {
         //获取菜单列表
-        List<Integer> roleList = ShiroKit.getUser().getRoleList();
+        List<Long> roleList = ShiroKit.getUser().getRoleList();
         if (roleList == null || roleList.size() == 0) {
             ShiroKit.getSubject().logout();
             model.addAttribute("tips", "该用户没有角色，无法登陆");
@@ -68,7 +68,7 @@ public class LoginController extends BaseController {
 
         //获取用户头像
         Long id = ShiroKit.getUser().getId();
-        User user = userRepository.findOne(id.intValue());
+        User user = userRepository.findOne(id);
         String avatar = user.getAvatar();
         model.addAttribute("avatar", avatar);
