+++ guns-api/src/main/java/cn/enilu/guns/api/controller/system/DeptContoller.java
@@ -56,7 +56,7 @@ public class DeptContoller extends BaseController {
     }
     @RequestMapping(method = RequestMethod.DELETE)
     @BussinessLog(value = "删除部门", key = "id", dict = DeptDict.class)
-    public Object remove(Integer id){
+    public Object remove(Long id){
         logger.info("id:{}",id);
         if (ToolUtil.isEmpty(id)) {
             throw new GunsException(BizExceptionEnum.REQUEST_NULL);
