+++ guns-dao/src/main/java/cn/enilu/guns/dao/system/DeptRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
  *
  * @author enilu
  */
-public interface DeptRepository extends PagingAndSortingRepository<Dept,Integer> {
+public interface DeptRepository extends PagingAndSortingRepository<Dept,Long> {
     List<Dept> findByPidsLike(String pid);
     @Query(nativeQuery = true,value = "SELECT id, pid AS pId, simplename AS NAME, ( CASE WHEN (pId = 0 OR pId IS NULL) THEN 'true' ELSE 'false' END ) AS isOpen FROM t_sys_dept")
     List tree();
