+++ guns-dao/src/main/java/cn/enilu/guns/dao/system/DictRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DictRepository extends JpaRepository<Dict, Integer> {
-    List<Dict> findByPid(Integer pid);
-    List<Dict> findByNameAndPid(String name,Integer pid);
+public interface DictRepository extends JpaRepository<Dict, Long> {
+    List<Dict> findByPid(Long pid);
+    List<Dict> findByNameAndPid(String name,Long pid);
     Dict findByName(String name);
 
     List<Dict> findByNameLike(String name);
