+++ guns-dao/src/main/java/cn/enilu/guns/dao/system/MenuRepository.java
@@ -27,7 +27,7 @@ public interface MenuRepository extends PagingAndSortingRepository<Menu,Long> {
             "CODE " +
             "INNER " +
             "JOIN ( SELECT ID FROM t_sys_menu WHERE ID IN ( SELECT menuid FROM t_sys_relation rela WHERE rela.roleid IN (?1))) m3 ON m1.id = m3.id WHERE m1.ismenu = 1 and m1.status = 1 ORDER BY levels, num ASC")
-    List getMenusByRoleIds(List<Integer> roleList);
+    List getMenusByRoleIds(List<Long> roleList);
     @Query(nativeQuery = true,value="SELECT m1.id AS id, m1.icon AS icon, ( CASE WHEN (m2.id = 0 OR m2.id IS NULL) " +
             "THEN 0 ELSE m2.id END ) AS parentId, m1. NAME AS NAME, m1.url AS url, m1.levels AS levels, m1.ismenu AS " +
             "ismenu, m1.num AS num, m1. CODE AS CODE,m1.status as status FROM t_sys_menu m1 LEFT JOIN t_sys_menu m2 " +
