+++ guns-dao/src/main/java/cn/enilu/guns/dao/system/UserRepository.java
@@ -10,6 +10,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
  *
  * @author enilu
  */
-public interface UserRepository extends JpaRepository<User,Integer>,JpaSpecificationExecutor<User> {
+public interface UserRepository extends JpaRepository<User,Long>,JpaSpecificationExecutor<User> {
     User findByAccount(String account);
 }
